Project in progress
The program will create P producer threads and one consumer thread which share access to a bounded buffer
of size B. The number of producers and the size of the bounded buffer will be available to the program as
command-line arguments.

 Each producer thread will process the transactions which are contained in the input file named "transN", where
N corresponds to the thread number (and thus is between 0 and 9). Each input file will contain zero or more lines,
where each line represents a transaction.